Improve hardening MODX instructions
Addresses some minor writing issues and improves the suggested rewrites to use the MODX error handler for NGINX (and NGINX 404 option, too). I don't have apache to test so might want to double-check the negative lookahead in the apache rewrite rule for dotfiles to allow `.well_known` for Let's Encrypt, but it did pass an online tester. Special thanks to Garry Nutting for the regex sanity checks and suggestions.