Changed structure of the demo-head and footerHTML
I think it's a really bad idea to have one chunk with only opening tags and the other one with closing tags that depend on each other. I'd avoid that in any situation. Editors like Ace will always show you if you missed a closing tag, you can expand/collapse those, as long as your HTML is valid and consistent. Having it the way it is now is sloppy and something we've seen in a lot of MODX sites that other people built. If it's in the official documentation, we can't even blame them. I'd recommend changing this for several obvious reasons. The most important reason would be to easily check if your HTML is valid. Having to touch two chunks when adding an opening or a closing tag that is wrapping the entire page is just a bad idea.